Move more things into testutils
