+++ commands/list_releases_test.go
@@ -6,14 +6,14 @@ import (
 	"os"
 	"testing"
 
-	"github.com/giantswarm/gsctl/config"
 	"github.com/giantswarm/gsctl/flags"
+	"github.com/giantswarm/gsctl/testutils"
 )
 
 // Test_ListReleases_Empty simulates the situation where there are no releases
 // (which is an exception)
 func Test_ListReleases_Empty(t *testing.T) {
-	dir, err := config.TempConfig("")
+	dir, err := testutils.TempConfig("")
 	if err != nil {
 		t.Error(err)
 	}
@@ -53,7 +53,7 @@ func Test_ListReleases_Empty(t *testing.T) {
 // Test_ListReleases_Nonempty simulates listing releases where several
 // items are returned.
 func Test_ListReleases_Nonempty(t *testing.T) {
-	dir, err := config.TempConfig("")
+	dir, err := testutils.TempConfig("")
 	if err != nil {
 		t.Error(err)
 	}
