+++ commands/logout_test.go
@@ -9,13 +9,13 @@ import (
 	"github.com/giantswarm/microerror"
 
 	"github.com/giantswarm/gsctl/client/clienterror"
-	"github.com/giantswarm/gsctl/config"
 	"github.com/giantswarm/gsctl/flags"
+	"github.com/giantswarm/gsctl/testutils"
 )
 
 // Test_LogoutValidToken tests the logout for a valid token
 func Test_LogoutValidToken(t *testing.T) {
-	dir, err := config.TempConfig("")
+	dir, err := testutils.TempConfig("")
 	if err != nil {
 		t.Error(err)
 	}
@@ -44,7 +44,7 @@ func Test_LogoutValidToken(t *testing.T) {
 
 // Test_LogoutInvalidToken tests the logout for an invalid token
 func Test_LogoutInvalidToken(t *testing.T) {
-	dir, err := config.TempConfig("")
+	dir, err := testutils.TempConfig("")
 	if err != nil {
 		t.Error(err)
 	}
@@ -78,7 +78,7 @@ func Test_LogoutInvalidToken(t *testing.T) {
 // Test_LogoutCommand simply calls the functions cobra would call,
 // with a temporary config path and mock server as endpoint.
 func Test_LogoutCommand(t *testing.T) {
-	dir, err := config.TempConfig("")
+	dir, err := testutils.TempConfig("")
 	if err != nil {
 		t.Error(err)
 	}
